Applying changes asked by the review
The different sources handles the URL creation based on the layer type
