+++ maplibre/src/stages/request_stage.rs
@@ -6,8 +6,10 @@ use crate::error::Error;
 use crate::io::source_client::{HttpSourceClient, SourceClient};
 use crate::io::tile_repository::TileRepository;
 use crate::io::TileRequest;
+use crate::io::TileRequestID;
 use crate::schedule::Stage;
 use crate::stages::SharedThreadState;
+use crate::style::source::TileAddressingScheme;
 use crate::{HttpClient, ScheduleMethod, Scheduler, Style};
 use std::collections::HashSet;
 
@@ -82,7 +84,11 @@ where
     SM: ScheduleMethod,
     HC: HttpClient,
 {
-    fn load(request_stage: &RequestStage<SM, HC>, coords: &WorldTileCoords, request_id: u32);
+    fn load(
+        request_stage: &RequestStage<SM, HC>,
+        coords: &WorldTileCoords,
+        request_id: TileRequestID,
+    );
 }
 
 pub struct TessellateSource;
@@ -92,9 +98,20 @@ where
     SM: ScheduleMethod,
     HC: HttpClient,
 {
-    fn load(request_stage: &RequestStage<SM, HC>, coords: &WorldTileCoords, request_id: u32) {
+    fn load(
+        request_stage: &RequestStage<SM, HC>,
+        coords: &WorldTileCoords,
+        request_id: TileRequestID,
+    ) {
         let client = SourceClient::Http(request_stage.http_source_client.clone());
         let coords = *coords;
+        let tile_coords = coords.into_tile(TileAddressingScheme::TMS).unwrap();
+        let url = format!(
+            "https://maps.tuerantuer.org/europe_germany/{z}/{x}/{y}.pbf",
+            x = tile_coords.x,
+            y = tile_coords.y,
+            z = tile_coords.z
+        );
 
         let state = request_stage.shared_thread_state.clone();
         request_stage
@@ -102,7 +119,7 @@ where
             .schedule_method()
             .schedule(Box::new(move || {
                 Box::pin(async move {
-                    match client.fetch(&coords).await {
+                    match client.fetch(&url).await {
                         Ok(data) => state
                             .process_tile(request_id, data.into_boxed_slice())
                             .unwrap(),
@@ -124,30 +141,20 @@ where
     SM: ScheduleMethod,
     HC: HttpClient,
 {
-    fn load(request_stage: &RequestStage<SM, HC>, coords: &WorldTileCoords, request_id: u32) {
-        // Sould have a RasterSourceClient or should fetching happen depending
-        // on the tile type ?
+    fn load(
+        request_stage: &RequestStage<SM, HC>,
+        coords: &WorldTileCoords,
+        request_id: TileRequestID,
+    ) {
         let client = SourceClient::Http(request_stage.http_source_client.clone());
         let coords = *coords;
-
-        let state = request_stage.shared_thread_state.clone();
-        request_stage
-            .scheduler
-            .schedule_method()
-            .schedule(Box::new(move || {
-                Box::pin(async move {
-                    match client.fetch(&coords).await {
-                        Ok(data) => state
-                            .process_tile(request_id, data.into_boxed_slice())
-                            .unwrap(),
-                        Err(e) => {
-                            log::error!("{:?}", &e);
-                            state.tile_unavailable(&coords, request_id).unwrap()
-                        }
-                    }
-                })
-            }))
-            .unwrap();
+        let tile_coords = coords.into_tile(TileAddressingScheme::XYZ).unwrap();
+        let url = format!(
+            "https://api.maptiler.com/tiles/satellite/{z}/{x}/{y}.jpg?key=qnePkfbGpMsLCi3KFBs3",
+            x = tile_coords.x,
+            y = tile_coords.y,
+            z = tile_coords.z
+        );
     }
 }
 
@@ -188,10 +195,7 @@ where
         coords: &WorldTileCoords,
         layers: &HashSet<String>,
     ) -> Result<bool, Error> {
-        // if !tile_repository.is_layers_missing(coords, layers) {
-        //     return Ok(false);
-        // }
-        let missed_sources = tile_repository.is_layers_missing(coords, layers);
+        let missed_sources = tile_repository.missed_layer(coords, layers);
         if missed_sources.is_empty() {
             return Ok(false);
         }
@@ -213,26 +217,6 @@ where
                     );
                 }*/
 
-                // let client = SourceClient::Http(self.http_source_client.clone());
-                // let coords = *coords;
-
-                // let state = self.shared_thread_state.clone();
-                // self.scheduler
-                //     .schedule_method()
-                //     .schedule(Box::new(move || {
-                //         Box::pin(async move {
-                //             match client.fetch(&coords).await {
-                //                 Ok(data) => state
-                //                     .process_tile(request_id, data.into_boxed_slice())
-                //                     .unwrap(),
-                //                 Err(e) => {
-                //                     log::error!("{:?}", &e);
-                //                     state.tile_unavailable(&coords, request_id).unwrap()
-                //                 }
-                //             }
-                //         })
-                //     }))
-                //     .unwrap();
                 let raster = "raster".to_string();
                 for source in missed_sources {
                     match source {
